rename junto -> flux where applicable
