+++ .github/workflows/release.yml
@@ -78,8 +78,8 @@ jobs:
       - name: Cache macOS release (x64)
         uses: actions/upload-artifact@v1
         with:
-          name: Junto-${{steps.ref.outputs.version}}.dmg
-          path: ./dist_electron/Junto-${{steps.ref.outputs.version}}.dmg
+          name: Flux-${{steps.ref.outputs.version}}.dmg
+          path: ./dist_electron/Flux-${{steps.ref.outputs.version}}.dmg
       - name: Cache latest-mac.yml
         uses: actions/upload-artifact@v1
         with:
@@ -124,8 +124,8 @@ jobs:
   #     - name: Cache nsis release
   #       uses: actions/upload-artifact@v1
   #       with:
-  #         name: Junto Setup ${{steps.ref.outputs.version}}.exe
-  #         path: ./dist_electron/Junto Setup ${{steps.ref.outputs.version}}.exe
+  #         name: Flux Setup ${{steps.ref.outputs.version}}.exe
+  #         path: ./dist_electron/Flux Setup ${{steps.ref.outputs.version}}.exe
   #     - name: Cache latest-windows.yml
   #       uses: actions/upload-artifact@v1
   #       with:
@@ -193,13 +193,13 @@ jobs:
       - name: Cache AppImage
         uses: actions/upload-artifact@v1
         with:
-          name: Junto-${{steps.ref.outputs.version}}.AppImage
-          path: ./dist_electron/Junto-${{steps.ref.outputs.version}}.AppImage
+          name: Flux-${{steps.ref.outputs.version}}.AppImage
+          path: ./dist_electron/Flux-${{steps.ref.outputs.version}}.AppImage
       - name: Cache .deb release (x64)
         uses: actions/upload-artifact@v1
         with:
-          name: junto_${{steps.ref.outputs.version}}_amd64.deb
-          path: ./dist_electron/junto_${{steps.ref.outputs.version}}_amd64.deb
+          name: flux_${{steps.ref.outputs.version}}_amd64.deb
+          path: ./dist_electron/flux_${{steps.ref.outputs.version}}_amd64.deb
       - name: Cache latest-linux.yml
         uses: actions/upload-artifact@v1
         with:
@@ -225,22 +225,22 @@ jobs:
       - name: Download AppImage
         uses: actions/download-artifact@v1
         with:
-          name: Junto-${{steps.ref.outputs.version}}.AppImage
+          name: Flux-${{steps.ref.outputs.version}}.AppImage
           path: .
       - name: Download deb
         uses: actions/download-artifact@v1
         with:
-          name: junto_${{steps.ref.outputs.version}}_amd64.deb
+          name: flux_${{steps.ref.outputs.version}}_amd64.deb
           path: .
       - name: Download macOS image
         uses: actions/download-artifact@v1
         with:
-          name: Junto-${{steps.ref.outputs.version}}.dmg
+          name: Flux-${{steps.ref.outputs.version}}.dmg
           path: .
       # - name: Download Windows Exe
       #   uses: actions/download-artifact@v1
       #   with:
-      #     name: Junto Setup ${{steps.ref.outputs.version}}.exe
+      #     name: Flux Setup ${{steps.ref.outputs.version}}.exe
       #     path: .
       # - name: Download latest-windows.yml
       #   uses: actions/download-artifact@v1
@@ -259,9 +259,9 @@ jobs:
           path: .
       - name: Generate SHA256 checksums
         run: |
-          sha256sum "junto_${{steps.ref.outputs.version}}_amd64.deb" >> "SHA256SUMS.txt"
-          sha256sum "Junto-${{steps.ref.outputs.version}}.AppImage" >> "SHA256SUMS.txt"
-          sha256sum "Junto-${{steps.ref.outputs.version}}.dmg" >> "SHA256SUMS.txt"
+          sha256sum "flux_${{steps.ref.outputs.version}}_amd64.deb" >> "SHA256SUMS.txt"
+          sha256sum "Flux-${{steps.ref.outputs.version}}.AppImage" >> "SHA256SUMS.txt"
+          sha256sum "Flux-${{steps.ref.outputs.version}}.dmg" >> "SHA256SUMS.txt"
           sha256sum "latest-linux.yml" >> "SHA256SUMS.txt"
           sha256sum "latest-mac.yml" >> "SHA256SUMS.txt"
       - name: Verify checksums
@@ -282,8 +282,8 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.github_token }}
         with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: ./junto_${{steps.ref.outputs.version}}_amd64.deb
-          asset_name: junto_${{steps.ref.outputs.version}}_amd64.deb
+          asset_path: ./flux_${{steps.ref.outputs.version}}_amd64.deb
+          asset_name: flux_${{steps.ref.outputs.version}}_amd64.deb
           asset_content_type: application/octet-stream
       - name: Upload AppImage binary (arm64)
         uses: actions/upload-release-asset@v1.0.1
@@ -291,8 +291,8 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.github_token }}
         with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: ./Junto-${{steps.ref.outputs.version}}.AppImage
-          asset_name: Junto-${{steps.ref.outputs.version}}.AppImage
+          asset_path: ./Flux-${{steps.ref.outputs.version}}.AppImage
+          asset_name: Flux-${{steps.ref.outputs.version}}.AppImage
           asset_content_type: application/octet-stream
       - name: Upload macOS image
         uses: actions/upload-release-asset@v1.0.1
@@ -300,8 +300,8 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: ./Junto-${{steps.ref.outputs.version}}.dmg
-          asset_name: Junto-${{steps.ref.outputs.version}}.dmg
+          asset_path: ./Flux-${{steps.ref.outputs.version}}.dmg
+          asset_name: Flux-${{steps.ref.outputs.version}}.dmg
           asset_content_type: application/octet-stream
       # - name: Upload Windows exe
       #   uses: actions/upload-release-asset@v1.0.1
@@ -309,8 +309,8 @@ jobs:
       #     GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       #   with:
       #     upload_url: ${{ steps.create_release.outputs.upload_url }}
-      #     asset_path: ./Junto Setup ${{steps.ref.outputs.version}}.exe
-      #     asset_name: Junto Setup ${{steps.ref.outputs.version}}.exe
+      #     asset_path: ./Flux Setup ${{steps.ref.outputs.version}}.exe
+      #     asset_name: Flux Setup ${{steps.ref.outputs.version}}.exe
       #     asset_content_type: application/octet-stream
       - name: Upload latest-linux.yml
         uses: actions/upload-release-asset@v1.0.1
