+++ scripts/clean-state.js
@@ -30,16 +30,16 @@ async function main() {
         process.env.HOME,
         "Library",
         "Application Support",
-        "junto"
+        "flux"
       );
       break;
     }
     case "win32": {
-      configPath = path.join(process.env.APPDATA, "junto");
+      configPath = path.join(process.env.APPDATA, "flux");
       break;
     }
     case "linux": {
-      configPath = path.join(process.env.HOME, ".config", "junto");
+      configPath = path.join(process.env.HOME, ".config", "flux");
       break;
     }
   }
