+++ src/constants/agents.ts
@@ -1,0 +1,5 @@
+export const JUNTO_AGENT =
+  "did:key:zQ3shpe32NARLJeuE3wwWLWLoy9NzkrD3oy1psPJNhbvkuMzK";
+
+export const AD4M_AGENT =
+  "did:key:zQ3shkkuZLvqeFgHdgZgFMUx8VGkgVWsLA83w2oekhZxoCW2n";
