+++ src/constants/config.ts
@@ -2,8 +2,3 @@ export const channelRefreshDurationMs = 5000;
 export const expressionGetRetries = 100;
 export const expressionGetDelayMs = 20;
 export const chatMessageRefreshDuration = 15000;
-
-export interface JuntoShortForm {
-  body: string;
-  background: [string];
-}
