+++ src/constants/languages.ts
@@ -9,9 +9,3 @@ export const PROFILE_EXPRESSION_OFFICIAL =
 
 export const SHORTFORM_EXPRESSION_OFFICIAL =
   "QmVDh4jd75BgPgiV6Lttjm7xQ8JQBkQJtqwBcPn4xZyXgP";
-
-export const JUNTO_AGENT =
-  "did:key:zQ3shpe32NARLJeuE3wwWLWLoy9NzkrD3oy1psPJNhbvkuMzK";
-
-export const AD4M_AGENT =
-  "did:key:zQ3shkkuZLvqeFgHdgZgFMUx8VGkgVWsLA83w2oekhZxoCW2n";
