+++ src/constants/types.ts
@@ -1,0 +1,4 @@
+export interface ShortForm {
+  body: string;
+  background: [string];
+}
