--- src/core/validate.ts
@@ -1,9 +1,0 @@
-import type { JuntoLanguageTypes } from "./juntoTypes";
-
-/// Should take an input bundle file, remove var DNA from bundle and then check hash of that bundle vs default installed Junto languages
-export default function validateLanguage(
-  bundle: string,
-  languageType: JuntoLanguageTypes
-): boolean {
-  return true;
-}
