+++ src/store/data/actions/getNeighbourhoodMetadata.ts
@@ -1,5 +1,5 @@
 import { print } from "graphql/language/printer";
-import { expressionGetRetries, expressionGetDelayMs } from "@/core/juntoTypes";
+import { expressionGetRetries, expressionGetDelayMs } from "@/constants/config";
 import { GET_EXPRESSION, PERSPECTIVE_LINK_QUERY } from "@/core/graphql_queries";
 import { LinkQuery } from "@perspect3vism/ad4m";
